// src/sites/item/lists/item/operations/item/richLongRunningOperationItemRequestBuilderPatchRequestConfiguration.ts
import type {RequestOption} from '@microsoft/kiota-abstractions';

export interface RichLongRunningOperationItemRequestBuilderPatchRequestConfiguration {
    /**
     * Request headers
     */
    headers?: Record<string, string[]> | undefined;
    /**
     * Request options
     */
    options?: RequestOption[] | undefined;
}